+++ src/main/java/org/example/vibelist/global/aop/UserLog.java
@@ -14,6 +14,6 @@ public class UserLog {
     private String domain; // ex: user,post, etc
     private String api;// ex: api/v1/post
     private String requestBody;
-    private Long duration;
+    private double duration;
     private LocalDateTime timestamp;
 }
